chore(ci): slightly shorten release workflow
Currently, this workflow lets PackSquash generate a ZIP with the pack,
and then adds three files to it that are checked in to the repository.

However, this is unnecessarily verbose: PackSquash has a feature to
force it to include files in the ZIP that it would otherwise skip.
This feature is already being used for the missing "icon.png" file, so
let's leverage it for the licensing and metadata files that are being
added too.